### 2025-05-30 f0d9314
citation style
