### 2025-05-30 #8317 @isaac-6
> ⚠️ [An error happened when generating the pdf](https://github.com/openjournals/joss-papers/actions/runs/15343841432?check_suite_focus=true). Problem with affiliations for Isaac Llorente-Saguer, perhaps the affiliations index need quoting?.

Fixed the indentation that caused the error.
