### 2025-05-30 #8317 @isaac-6
Perhaps **ellessenne** could be a good reviewer/editor candidate, given their published R package for a biological measure (comorbidity scores).

Also **nhejazi**, given their published R package for biomarker discovery.
