### 2025-06-02 5ce38bd
Paper (#15)
* removed project-specific files

* Action: paper draft

* removed legacy function

* Updated installation line to point to the branch.

* format fix

* citation style

* added "JOSS submitted" badge
