### 2025-06-09 79e1c43
Changed default parallel to TRUE, and monitor to FALSE, to run experiments.
