### 2025-06-10 3ac5791
Enabled user-defined literature biomarkers, or to utilise those included in the package.
