### 2025-06-17 5fbf6f1
Parallel defaults to FALSE (uses 1  CPU core). Change it to TRUE to use all available cores, or to a specific number of cores (e.g., 8).
