+++ paper/paper.bib
@@ -1,4 +1,4 @@
-@article{llorente:2024,
+@article{llorente2024,
   title={A data-driven framework for biomarker discovery applied to optimizing modern clinical and preclinical trials on Alzheimer’s disease},
   author={Llorente-Saguer, Isaac and Oxtoby, Neil P},
   journal={Brain Communications},
