+++ paper/paper.bib
@@ -8,3 +8,32 @@
   year={2024},
   publisher={Oxford University Press UK}
 }
+
+@article{scrucca2016some,
+  title={On some extensions to GA package: hybrid optimisation, parallelisation and islands evolution},
+  author={Scrucca, Luca},
+  journal={arXiv preprint arXiv:1605.01931},
+  year={2016}
+}
+
+@Article{lme4,
+  title = {Fitting Linear Mixed-Effects Models Using {lme4}},
+  author = {Douglas Bates and Martin M{\"a}chler and Ben Bolker and Steve Walker},
+  journal = {Journal of Statistical Software},
+  year = {2015},
+  volume = {67},
+  number = {1},
+  pages = {1--48},
+  doi = {10.18637/jss.v067.i01}
+  
+@Article{longpower,
+  title = {Power and Sample Size for Longitudinal Models in R-The longpower Package and Shiny App},
+  author = {Samuel Iddi and Michael C. Donohue},
+  journal = {R Journal},
+  year = {2022},
+  volume = {14},
+  number = {1},
+  pages = {264-281},
+  doi = {10.32614/RJ-2022-022},
+  publisher = {R Foundation for Statistical Computing},
+}
