### 2025-09-15 a53a341
added package references to the paper
