### 2025-09-15 ed48863
Added seminal CVR concept paper
