+++ paper/paper.bib
@@ -9,6 +9,17 @@
   publisher={Oxford University Press UK}
 }
 
+@article{saguer2022composite,
+  title={Composite SUVR: A new method for boosting Alzheimer’s disease monitoring and diagnostic performance, applied to tau PET},
+  author={Saguer, Isaac Llorente and Busche, Marc A and Oxtoby, Neil P},
+  journal={Alzheimer's \& Dementia},
+  volume={18},
+  pages={e063177},
+  year={2022},
+  publisher={Wiley Online Library}
+}
+
+
 @article{scrucca2016some,
   title={On some extensions to GA package: hybrid optimisation, parallelisation and islands evolution},
   author={Scrucca, Luca},
@@ -16,7 +27,7 @@
   year={2016}
 }
 
-@Article{lme4,
+@article{lme4,
   title = {Fitting Linear Mixed-Effects Models Using {lme4}},
   author = {Douglas Bates and Martin M{\"a}chler and Ben Bolker and Steve Walker},
   journal = {Journal of Statistical Software},
@@ -26,7 +37,7 @@
   pages = {1--48},
   doi = {10.18637/jss.v067.i01}
   
-@Article{longpower,
+@article{longpower,
   title = {Power and Sample Size for Longitudinal Models in R-The longpower Package and Shiny App},
   author = {Samuel Iddi and Michael C. Donohue},
   journal = {R Journal},
