### 2025-09-15 2263ee9
fixed bibliography style
