+++ paper/paper.bib
@@ -36,6 +36,7 @@
   number = {1},
   pages = {1--48},
   doi = {10.18637/jss.v067.i01}
+}
   
 @article{longpower,
   title = {Power and Sample Size for Longitudinal Models in R-The longpower Package and Shiny App},
