+++ paper/paper.bib
@@ -6,7 +6,8 @@
   number={6},
   pages={fcae438},
   year={2024},
-  publisher={Oxford University Press UK}
+  publisher={Oxford University Press UK},
+  doi={10.1093/braincomms/fcae438}
 }
 
 @article{saguer2022composite,
@@ -16,7 +17,8 @@
   volume={18},
   pages={e063177},
   year={2022},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/alz.063177}
 }
 
 
@@ -24,7 +26,8 @@
   title={On some extensions to GA package: hybrid optimisation, parallelisation and islands evolution},
   author={Scrucca, Luca},
   journal={arXiv preprint arXiv:1605.01931},
-  year={2016}
+  year={2016},
+  doi={10.48550/arXiv.1605.01931}
 }
 
 @article{lme4,
