### 2025-09-22 #8317 @donishadsmith
Hi @jromanowska

I would like to volunteer to review biodiscvr! After reviewing the paper, I noted that this package appears to be image agnostic. So, I would be very interested in testing its application on biomarkers from fmri data.
