### 2025-09-22 #8317 @isaac-6
@donishadsmith That would be great! Let me know if you need any help adapting to your data. The framework is flexible, so if there is any knowledge-specific information, it can be adapted. The important part is to define the metric to improve. In my case it's a composite metric, mostly centred on reducing the sample size estimate for a clinical trial, but it can very well be AUC for classification, etc.
