### 2025-09-24 #9070 @editorialbot
[REVIEW]: biodiscvr: Biomarker Discovery Using Composite Value Ratios
**Submitting author:** @isaac-6 (<a href="http://orcid.org/0000-0002-7612-5331">Isaac Llorente-Saguer</a>)
**Repository:** https://github.com/isaac-6/biodiscvr
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0.0
