### 2025-10-07 #9070 @donishadsmith
Hi @jromanowska! I apologize for the delay @isaac-6. I recently accepted a postdoc position so was in the process of moving and am currently in the process of onboarding. I checked of some additional things that I could now though I wanted to request one week for me to return to the review.

@jromanowska I also have a quick question about the license, this package is licensed under GPL-3.0; however,
GPL-3.0-only is the short identifier listed as an OSI-approved license. In this case, can this license still be accepted only would it need to be changed? Thanks!
