### 2025-10-07 f6b82f0
specified license to GPL-3.0-or-later
