### 2025-10-07 #9070 @isaac-6
@jromanowska @donishadsmith No modifications were made. I have now specified GPL-3.0-or-later; thank you for the feedback!

Note that the branch I submitted for JOSS is the one called "paper".

PS: Congratulations on the postdoc position, Donisha!
