### 2025-10-07 e893672
specified license as GPL-3.0-or-later
