### 2025-10-21 isaac-6/biodiscvr#16 @donishadsmith
[JOSS Review] Generalizability of the current implementation of the `biodiscvr`.
Hi @isaac-6,

This issue is to inquire about the intended generalizability of `biodiscvr` for openjournals/joss-reviews#9070.
The "Summary" and "Statement of need" sections of the JOSS paper suggest that the software is broadly applicable and could be adapted to different domains. However, a deeper look at the paper, source code, and documentation reveals that the current implementation seems specifically tailored for Alzheimer's disease research using PET imaging.

For example:
- In one line of the "Mathematics" section of the paper uses domain-specific term ("amyloid-positive").
- The source code requires specific columns such as "DX" (cognitive status) and "AB" (amyloid positivity), with hardcoded checks for groups such as "CU" and "CI". Certain functions are programmed to either stop run execution, perform early "NULL" returns, skip subject filtering steps, or return "NA" values if these required are absent from "data.csv". Just to be sure, I ran the example in the vignette for synthetic data, which runs perfectly when all expected information is present and produces the previously listed issues when "data.csv" is modified to remove expected columns or change the data type (i.e. binary vector for "DX").
- The code also contains hardcoded value constraints for specific brain regions (e.g., entorhinal, amygdala) based on literature for its intended domain.

This Alzheimer-specific implementation is clear in the source code, the vignette, and function documentation (such as `create_demographics`). However, this specificity is not immediately apparent in the main "README.md" file (other than a reference to "preparing SUV columns" in one line) or the introductory sections of the JOSS paper.

To avoid confusion for users from other domains, I suggest making this focus on Alzheimer's research more explicit in both the "README.md" file and the JOSS paper. If there are plans to make the software more domain-agnostic in the future, it would be helpful to mention this in the paper as well.

Also, another note, in the statement of need for the JOSS paper, there was a mention of swapping discovery algorithms; however, I have only seen the genetic algorithm be used in the vignette example and the source code unless I misunderstood something.

Thank you for your work on this project!
