### 2025-10-21 isaac-6/biodiscvr#17 @donishadsmith
[JOSS Review] Add tests
For openjournals/joss-reviews#9070.

Currently, the repo for the package has a GitHub workflow that performs a `R CMD check` to ensure the package and its dependencies can be installed on the latest available R (4.5.1), including some other R quality control tests. However, one `R CMD check` that is missing is the testing check using a package such as [testthat](https://testthat.r-lib.org/) ([here is an example of testhat from the tidyr package](https://github.com/tidyverse/tidyr/tree/main/tests)) I also don't see any manual tests that are implemented outside of a package.

While the current checks are robust against installation issues, it won't protect against changes in dependencies that may result in functions not running or code changes that could cause silent errors. If possible, could you add some basic tests, to your package to ensure functionality. These tests could simply use the synthetic data available in the repo and use parameter settings to get the tests to run as quickly as possible. These tests don't have to intensive checks for the sake of this review; they can perform a basic check to ensure functions produce the output of a specific class or don't produce an error.

Also, adding a note in the README about which versions of R your package is compatible with (or the version you want to support, you could just state something along the line of "intended for R version 4.5.1 and above") would also be helpful.
