### 2025-10-21 isaac-6/biodiscvr#18 @donishadsmith
[JOSS Review] Contributing Guidelines
For openjournals/joss-reviews#9070.

This issue is a pretty easy one. Currently, the [contributing guideline](https://github.com/isaac-6/biodiscvr?tab=readme-ov-file#contributing) reference the code of conduct, which is great! I believe adding a sentence similar to the "Support and Contributions" section of this recently published [JOSS package](https://github.com/uislambekov/TDAvec?tab=readme-ov-file#support-and-contributions) would be good for users to know where suggestions, requests, and issues should go.
