### 2025-11-17 isaac-6/biodiscvr#16 @isaac-6
Thank you for raising these questions, @donishadsmith . It's a fair point. While the concept is agnostic to such specific variables and domain, I did tailor the actual working code to my specific problem (Alzheimer's disease pathology image-based monitoring). I believe this is highly adaptable, as I have personally used it in three different specific problems and data sources (tau PET, MRI, and amyloid peptides). However, I understand that such applications currently require some code adaptation, so I am happy to reword the summary and statement of need.

Would that work?
