### 2025-11-17 b6e455f
docs: added GitHub Issues link for questions, bug reports, and feature suggestions
