### 2025-12-16 fdf874d
Add automated test suite for JOSS review (local #17)
