+++ inst/files/config_synth.yaml
@@ -66,8 +66,8 @@ power_params:
 # Settings for the GA run by get_individual_fitness.
 # These are examples; adjust based on your specific GA implementation and needs.
 genetic_algorithm:
-  popSize: 32           # Population size per generation. Time scales linearly.
-  maxiter: 30          # Maximum number of generations (iterations) default 600. Time scales linearly.
+  popSize: 16           # Population size per generation (e.g., 32). Time scales linearly. 
+  maxiter: 10          # Maximum number of generations (iterations) default 600. Time scales linearly.
   run: 15              # Early stopping if no fitness improvement in X generations default 100
   pmutation: 0.5        # Probability of mutation
   crossover: "gareal_blxCrossover"
