+++ DESCRIPTION
@@ -41,5 +41,7 @@ Suggests:
     parallel,
     doParallel,
     knitr,
-    rmarkdown
+    rmarkdown,
+    testthat (>= 3.0.0)
 VignetteBuilder: knitr
+Config/testthat/edition: 3
