### 2025-12-16 isaac-6/biodiscvr#17 @isaac-6
Hi Donisha,

I have addressed your feedback regarding automated testing and README documentation:

- Automated Testing: Implemented a testthat suite (tests/testthat/) that verifies the full workflow (loading, preprocessing, demographics, and discovery) using the package's synthetic data.

- README Update: Added a compatibility note stating the package is intended for R version 4.5.1 and above.

The package now passes R CMD check with all tests active. Thank you for the suggestions!
