### 2025-12-22 9cfb264
docs: add disclaimer that while framework is conceptually general, current implementation is AD‑specific
