### 2026-01-05 #9070 @jromanowska
👋 @nhejazi & @edeno, I permit myself to ask you once more about your availability for reviewing this for JOSS. One of the reviewers has almost completed the checklist, while another one couldn't do it due to other tasks, so I need someone for a second opinion.
