### 2026-01-22 #9070 @jromanowska
@isaac-6, I have some comments to the paper:

- **page 1, line 17:** why is there "e.g., in config.yaml"? Are there other ways of defining inclusion criteria?
- **page 1, lines 26-27:** I'd suggest "large number of possible feature groupings" instead of "combinatorial explosion of possible feature groupings"
- **page 1, lines 27-29:** you're mentioning "encoding" and "applying exploratory algorithm" - could you be more specific? Any references here?
- **page 1, line 30:** "core methodology has been previously described" - also missing reference?
- **page 1, lines 34-38:** this is basically the same as in the summary and the sentence above - reword or delete this paragraph.
- **pages 1-2, lines 39-43:** please, be more concrete here; what is "multicohort regularisation", how is the algorithm making use of "multiple independent dataset without requiring data merging"? Some references here?
- **page 2, line 48:** what is "Pareto frontier"? (If it's a term obvious for researchers in neuroimaging, then you can skip this question)
- **"Mathematics" section:** I couldn't understand this section. Perhaps you'd need to rephrase it somehow or include other information. I guess the main question is: whom is this paper designed for? Is it the mathematicians that want to understand your method or is it the neuroscientists or bioinformaticians who would like to use your method?

Generally, in "Statement of need", I'd like to see more of refering to other methods or similar packages used for simpler tasks(?) and a bit of exploration of the "gap" you're trying to fill in. It's good that you mention which packages are you using in the implementation, but you're not mentioning _how_.

There could be also more information about the general flow of the tasks that your package allows to do. Right now, you're naming a lot of things one can do with the package (lines 17-19), but I didn't get any more information about the implementation from reading the paper.
