### 2026-01-22 #9070 @jromanowska
@isaac-6, there are two more things I'd like you to improve:

- [x] **work a bit more on the vignettes**. There is basically no explanation of each step and no results shown, no walk-through of what is going on in each step. The code is useful, but there should be more text. Especially, since your paper states that it is a framework that was developed with AD as an example, but can be adapted to other outcomes.
- [x] Moreover, you do not state how one can adapt or extend the code - **which functions and files should be changed?** This should be part of "Contributing" section in your README.
