### 2026-02-04 cf18600
docs: revision for JOSS; address editor comments. (v1.2.1)
- Restructured paper: merged repetitive sections and clarified inclusion criteria.
- Expanded paper: added 'Software Architecture and Workflow' section.
- Refined 'Methods': formalised multicohort regularisation logic and objective functions.
- Updated README: added developer guide for extending/adapting the framework.
