+++ paper/paper.bib
@@ -52,3 +52,23 @@
   doi = {10.32614/RJ-2022-022},
   publisher = {R Foundation for Statistical Computing},
 }
+
+@article{llorente2026enhanced,
+  title={Enhanced monitoring of Alzheimer's disease brain atrophy using composite value ratios of volumes},
+  author={Llorente-Saguer, Isaac and Oxtoby, Neil P},
+  journal={Brain Communications},
+  volume={8},
+  number={1},
+  pages={fcaf497},
+  year={2026},
+  publisher={Oxford University Press UK}
+}
+
+@article{llorente2025theta,
+  title={Theta, a multidimensional ratio biomarker applied to five amyloid beta peptides for investigations in familial Alzheimer’s disease},
+  author={Llorente-Saguer, Isaac and Arber, Charles and Oxtoby, Neil P},
+  journal={medRxiv},
+  pages={2025--08},
+  year={2025},
+  publisher={Cold Spring Harbor Laboratory Press}
+}
