+++ paper/paper.md
@@ -34,13 +34,13 @@ utilities, expected variable names, and default assumptions correspond to
 common AD datasets (e.g., cognitive status groupings, amyloid/tau measures,
 and region‑of‑interest conventions).
 
-The core functionality utilizes a Genetic Algorithm (GA) to search the
+The core functionality utilises a Genetic Algorithm (GA) to search the
 feature space for optimal numerator and denominator combinations based on
 biomarker performance metrics calculated using linear mixed‑effects models
 (Group Separation, Sample Size Estimates).
 
-The framework allows users to define inclusion criteria (e.g., in
-`config.yaml`), preprocess data, run the discovery workflow, perform
+The framework allows users to define inclusion criteria (in
+`config.yaml` or overriding workflow functions), preprocess data, run the discovery workflow, perform
 regional ablation analyses, and evaluate lists of biomarkers (including
 those discovered by the algorithm) across multiple datasets. Although the
 current release is AD‑focused, the architecture is designed to support
@@ -49,66 +49,120 @@ future extensions toward more domain‑agnostic workflows.
   
 # Statement of need
 
-A **composite value ratio (CVR)** [@saguer2022composite] is defined as the
-ratio between two composite aggregations of features. This concept is
-particularly powerful in neuroimaging, where shared confounding factors
-across features can cancel out, revealing more robust biomarkers.
-However, brute‑force exploration of CVRs is computationally infeasible due
-to the combinatorial explosion of possible feature groupings. Encoding this
-search space and applying an exploratory algorithm enables the discovery of
-high‑performing biomarkers tailored to specific clinical or research goals.
+A **composite value ratio (CVR)** [@saguer2022composite] is defined as the ratio 
+between two composite aggregations of features. This concept is widely used in 
+neuroimaging, where shared confounding factors (such as global signal intensity 
+or individual variations in tracer pharmacokinetics) can partially cancel out 
+when using ratios, thereby revealing more robust biological signals. 
+Traditionally, these ratios (e.g., the Standardised Uptake Value ratio in PET) 
+are determined based on expert knowledge and manual selection of target 
+and reference regions.
 
-While the core methodology has been previously described, there is currently no 
-openly available framework that implements it in a modular and extensible way. 
-This package fills that gap by providing a flexible implementation that allows 
-users to swap out discovery algorithms, redefine fitness metrics, and adapt the 
-pipeline to different domains.
+While standard feature selection frameworks (e.g., those implementing Lasso or 
+Elastic Net regularisation) are effective for identifying individual predictors 
+in linear models, they are not designed to explore the combinatorial and symbolic 
+space of feature groupings and ratios. Exhaustive search of all possible CVR 
+combinations is computationally unfeasible; for a typical neuroimaging template 
+with 100 regions (or features), the search space for possible ratios exceeds $10^{46}$ combinations. 
+`biodiscvr` addresses this by providing an automated exploratory framework 
+to discover high-performing CVRs tailored to specific clinical or research objectives, 
+such as maximising group separation or minimising sample size requirements [@saguer2022composite, @llorente2024, @llorente2026enhanced].
 
-Because the package was developed in the context of Alzheimer’s disease
-research, several components—such as expected column names, group
-definitions, and domain‑specific constraints—reflect this initial focus.
-Users applying *biodiscvr* to other domains may need to adapt these
-elements, but the underlying framework is general and can operate on
-arbitrary numeric features.
-
-A key innovation in this framework is the integration of multicohort
-regularisation, which enhances generalisability by allowing the algorithm
-to be informed by multiple independent datasets without requiring data
[…]
