+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: biodiscvr
 Title: Biomarker Discovery Using Composite Value Ratios
-Version: 1.2.0
+Version: 1.2.1
 Date: 2025
 Authors@R: 
     person("Isaac", "Llorente-Saguer", , "illorentes@gmail.com", role = c("aut", "cre"), comment = c(ORCID = "0000-0002-7612-5331"))
