### 2026-02-04 #9070 @isaac-6
Thank you again, @jromanowska, for the constructive feedback. In response, we have revised both the manuscript and the repository to improve clarity, expand the description of the software architecture, and provide clearer guidance for future users and developers.
Detailed responses to each point are provided below.

1. (page 1, line 17)
Comment: Why is there "e.g., in config.yaml"? Are there other ways of defining inclusion criteria?
Response: We have clarified that inclusion criteria can be defined in the config.yaml file or by overriding specific workflow functions within the R environment.

2. (page 1, lines 26-27)
Comment: I'd suggest "large number of possible feature groupings" instead of "combinatorial explosion."
Response: Thank you for the suggestion; we have updated the text stating “Exhaustive search of all possible CVR combinations is computationally unfeasible”.

3. (page 1, lines 27-30)
Comment: Could you be more specific about "encoding" and "exploratory algorithm"? Missing references?
Response: We have updated this section to explicitly mention the use of vector-based encoding for the search space and the application of a Genetic Algorithm (GA) as the exploratory engine. We have added references to the foundational CVR theory [@saguer2022composite] and the subsequent application studies [@llorente2024, @llorente2026enhanced].

4. (page 1, lines 34-38)
Comment: Reword or delete the paragraph that repeats the summary.
Response: We have merged these sections into a single, cohesive paragraph. It now explicitly states the gap the package fills: providing a modular, extensible, and domain-agnostic framework for CVR discovery that moves beyond the specific Alzheimer’s disease (AD) templates included as defaults.

5. (pages 1-2, lines 39-43)
Comment: Be more concrete about how the algorithm uses multiple datasets without merging. References?
Response: We have significantly expanded the "Statement of Need" and "Methods" sections to explain this. The software evaluates candidate biomarkers within each cohort independently to preserve cohort-specific structures and avoid batch-effect contamination. It then aggregates these results using a "consensus fitness" score that penalizes inconsistent performance (regularisation). As this implementation of multicohort consensus is an original contribution of this software, we have detailed the mathematical logic in the revised Methods section.

6. (page 2, line 48)
Comment: What is "Pareto frontier"?
Response: We have added a brief definition: the set of solutions where one objective (e.g., length) cannot be improved without degrading another (e.g., height).

7.
Comment: Rephrase the Mathematics section. Whom is this paper designed for?
Response: We have fully restructured the former ‘Mathematics’ section into a clearer ‘Methods’ section. It now progresses logically from individual statistical models to the global optimisation strategy. The paper is intended for bioinformaticians and research software engineers seeking a reproducible approach to biomarker discovery. By detailing the objective function and consensus logic, we provide sufficient depth for methodologists while keeping the overall workflow accessible to clinical researchers.

8.
Comment: Need more information about the general flow and implementation.
Response: We have added a new section titled "Software Architecture and Workflow." This section describes the four layers of the package: Configuration, Data Harmonisation, the Discovery Engine, and Parsimony Refinement. This clarifies how the various R packages (GA, lme4, lmmpower) are integrated into the pipeline.

9. Repository Improvements (Vignettes and Extensibility)
Comment: Work on vignettes and state how to adapt or extend the code.
Response: We have made the following improvements to the repository:
•	Vignettes: [to do]
•	Extensibility: We have added a dedicated "Adapting to Other Research Domains" subsection to the README.md under “Contributing”. This provides specific guidance on which files to modify to adapt the framework to new clinical domains or custom fitness metrics.
