### 2026-02-05 #9070 @jromanowska
@isaac-6 Thank you for the extensive explanation above and all the improvements!
I still need to re-read the paper, but just two questions for now:

- you write that the vignettes are still "[to do]" - should I wait?
- could you point to the PR or commit for the change(s) in point 9. above?

Thanks for the great work!
