### 2026-02-05 #9070 @isaac-6
@jromanowska Thank you for the kind comments. The changes related to points 1-9 were implemented in commit cf18600 (link below):
[https://github.com/isaac-6/biodiscvr/commit/cf1860092b1db6191221f0799494f5b561835b52](https://github.com/isaac-6/biodiscvr/commit/cf1860092b1db6191221f0799494f5b561835b52)

I plan to improve the vignettes today; tomorrow at the latest.
