### 2026-02-05 ac404b8
Added some guidance on the vignettes.
