### 2026-02-05 #9070 @isaac-6
@jromanowska Vignette changes have been implemented in the following: [commit](https://github.com/isaac-6/biodiscvr/commit/ac404b83351ba04a92540b783bb74f3f0c0b61c0). Note that I have changed the multicohort discovery evaluation to TRUE, so running the examples will take a few extra seconds.

**Revisions to the vignettes**

- _Domain adaptability_: The biodiscvr vignette now includes a new section, “Adapting the Framework to Other Domains,” explaining how to apply the method in other fields by adjusting features and outcomes.
- _Walk‑through and interpretation_: The biodiscvr_synth vignette now has a section called “Interpreting Discovery Results,” which walks through the key outputs and how to evaluate a discovered biomarker.
- _Demonstrating results_: The synthetic‑data vignette now shows the output of a discovery run using knitr::kable, addressing the earlier point that no results were displayed.

I hope these updates make the package easier to understand and use. Happy to revise further based on your feedback.

Cheers,
Isaac
