### 2026-02-05 #9070 @jromanowska
@isaac-6 - the `biodiscvr_synth` vignette is much better now, but I don't understand why you use so extensively comments in the code chunks, instead of actually writing the text as a narrative for guiding the user through the code. Moreover, none of the output is shown, you only explain what the output should look like. The vignettes would be much better if the output was under the code, where it belongs. Could you improve the vignettes accordingly?

Also: this is not a must, but {pckdown} and GitHub Pages are very easily configurable to create a pretty documentation webpage out of your .Rd files and vignettes, so you might consider that when you plan your next improvements in the future (no need to do it now, before acceptance). :)
