### 2026-02-05 64766df
added missing DOIs
