+++ paper/paper.bib
@@ -61,7 +61,8 @@
   number={1},
   pages={fcaf497},
   year={2026},
-  publisher={Oxford University Press UK}
+  publisher={Oxford University Press UK},
+  doi={10.1093/braincomms/fcaf497}
 }
 
 @article{llorente2025theta,
@@ -70,5 +71,6 @@
   journal={medRxiv},
   pages={2025--08},
   year={2025},
-  publisher={Cold Spring Harbor Laboratory Press}
+  publisher={Cold Spring Harbor Laboratory Press},
+  doi={10.1101/2025.08.06.25333131}
 }
