### 2026-02-05 #9070 @isaac-6
Hi @jromanowska , thank you again for the detailed and helpful feedback! I’ve added the missing DOIs [(commit link)](https://github.com/isaac-6/biodiscvr/commit/64766df851bdda5707538efccfc98a6901c2272e).

Regarding the commented code lines: I tend to annotate code because it helps me understand the logic when revisiting it after some time, and I thought these comments might also benefit users who want to follow the internal steps more closely. The vignette introduction already provides a full workflow overview, so the inline comments are meant as an additional layer of clarity. That said, I’ll expand the explanation of the steps to make this more explicit.

I’ll also look into why the table renders correctly when I knit the vignettes locally but not on your end.
