### 2026-02-06 #9070 @jromanowska
@isaac-6, maybe I was not clear enough. A vignette is like a case demo - explaining what you can do with the package and what kind of output one gets from different functions/code chunks. That is why one writes it in Rmarkdown - this format allows for interlacing code, output, and commentary. Your code with comments is very useful, but I'd see it fit in a script (or more smaller scripts), not in an Rmd file. So for me, it looks like you mixed scripting with writing vignettes. Look e.g., [at this vignette](https://bioconductor.org/packages/release/bioc/vignettes/kmcut/inst/doc/kmcut_intro.html) to see what I mean.
