### 2026-02-10 eaf7c0b
Improved vignettes (explanations, style).
