+++ vignettes/biodiscvr_synth.Rmd
@@ -1,397 +1,163 @@
 ---
-title: "Biodiscvr: Synthetic Data Workflow Example"
-output: rmarkdown::html_vignette
+title: "Biodiscvr: Synthetic Data Case Demo"
+date: "`r Sys.Date()`"
+toc-title: "Overview"
+output: 
+  rmarkdown::html_vignette:
+    toc: true
+    number_sections: true
 vignette: >
-  %\VignetteIndexEntry{Biodiscvr: Synthetic Data Workflow}
+  %\VignetteIndexEntry{Biodiscvr: Synthetic Data Case Demo}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
+
 
 ```{r setup, include = FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>",
-  eval = TRUE, # Default to evaluating chunks for this synthetic example
-  fig.width = 7,
-  fig.height = 5,
-  warning = FALSE, 
-  message = FALSE 
+  eval = TRUE,
+  message = FALSE,
+  warning = FALSE
 )
-# Load necessary packages
-# Ensure these are in Suggests or Imports in DESCRIPTION
 library(biodiscvr)
-library(yaml)
-library(dplyr)
-library(readr)
 library(knitr)
+library(dplyr)
+library(yaml)
 ```
 
-## Introduction
+# Introduction
 
-This vignette demonstrates the end-to-end workflow for using the `biodiscvr` package. We will use a synthetic dataset included with the package to illustrate all major steps, from loading data to discovering and evaluating biomarkers.
+The `biodiscvr` package provides a data-driven framework for the discovery of optimized biomarkers. Its showcase focuses on identifying **Composite Value Ratios (CVRs)**: combinations of features mathematically optimized to maximize clinical group separation while minimizing the sample size required for clinical trials.
 
-This allows running the entire pipeline and understand the inputs, outputs, and capabilities of `biodiscvr`.
+This vignette serves as a case demonstration using synthetic multi-cohort data.
 
-The workflow covers:
+# Environment Setup
 
-1.  Setting up paths to the (synthetic) data and output locations.
-2.  Loading the synthetic datasets using `load_datasets()`.
-3.  Checking data integrity, preparing SUV data, calculating time variables, and applying preprocessing filters using the consolidated `preprocess_data()` function.
-4.  Generating demographic summaries using `create_demographics_table()`.
-5.  Running a single-cohort biomarker discovery (`biodiscvr_single`).
-6.  Running a multi-cohort biomarker discovery (`biodiscvr_multicohort`).
-7.  Evaluating the biomarkers found by the multi-cohort discovery across all available synthetic cohorts (`evaluate_biomarkers()`).
-8.  Optionally, refining a top biomarker using iterative ablation (`refine_biomarker_by_ablation`).
+We start by locating the synthetic data and example configuration files bundled with the package, and defining a temporary output directory.
 
-## Setup: Synthetic Data and Output Paths
+```{r setup_paths}
+# Locate package-internal synthetic data and configs
+synth_data_dir <- system.file("synthdata", package = "biodiscvr", mustWork = TRUE)
+pkg_files      <- system.file("files", package = "biodiscvr", mustWork = TRUE)
 
-The `biodiscvr` package includes synthetic data in `inst/synthdata/` (cohort1, cohort2, cohort3) and example configuration files in `inst/files/` tailored for this synthetic data.
-Code is set to eval=FALSE. Change to TRUE or run it manually on a script.
-
-```{r define_paths_synth_vignette, eval=TRUE}
[…]
