+++ vignettes/biodiscvr.Rmd
@@ -1,6 +1,11 @@
 ---
 title: "Biodiscvr Workflow with User Data"
-output: rmarkdown::html_vignette
+date: "`r Sys.Date()`"
+toc-title: "Overview"
+output: 
+  rmarkdown::html_vignette:
+    toc: true
+    number_sections: true
 vignette: >
   %\VignetteIndexEntry{Biodiscvr Workflow with User Data}
   %\VignetteEngine{knitr::rmarkdown}
@@ -11,16 +16,18 @@ vignette: >
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>",
-  eval = FALSE # <<< IMPORTANT: Default to NOT evaluating chunks
+  eval = FALSE
 )
 # Load necessary packages for code display and explanation
 library(biodiscvr)
-library(yaml) # Needed for reading config examples potentially
+library(yaml)
 ```
 
 ## Introduction
+**Quick Start Note**: For a rapid, executed demonstration using pre-bundled synthetic data, please see the Case Demo: 
+`vignette("biodiscvr_synth", package = "biodiscvr")`
 
-This vignette demonstrates the typical end-to-end workflow for using the `biodiscvr` package to discover and evaluate biomarkers using **your own data**.
+The current vignette demonstrates the typical end-to-end workflow for using the `biodiscvr` package to discover and evaluate biomarkers using **your own data**.  
 
 We will cover the following steps:
 
@@ -32,9 +39,7 @@ We will cover the following steps:
 6.  Evaluating predefined literature biomarkers using `evaluate_literature_biomarkers()`.
 7.  Identifying key output files for reporting.
 
-**Note:** Because this vignette uses placeholders for user-specific file paths, the R code chunks are generally set to `eval=FALSE`. You will need to copy, paste, and adapt the code chunks into your own R script or console, replacing the placeholder paths with the actual locations of your files.
-
-#### **Adapting the Framework to Other Domains**
+### **Adapting the Framework to Other Domains**
 While this vignette uses Alzheimer's Disease (AD) examples (e.g., amyloid status and neuroimaging regions), `biodiscvr` is designed to be a general-purpose biomarker discovery engine. To adapt this to a different field (such as Oncology or astronomy):
 
 1.  **Define your Features:** Instead of brain regions, provide your domain-specific numeric features (e.g., gene expression levels, spectral intensities, imaging-derived metrics).
@@ -47,6 +52,18 @@ For technical instructions on how to modify the source code or fitness metrics f
 
 
 ## Setup: User Files and Paths
+
+BioDisCVR uses a decentralized loading system. Each cohort should be stored in its own folder within a main root directory. A typical setup looks like this:
+
+```text
+my_data_root/
+├── Cohort_A/
+│   ├── data.csv            # Metadata
+│   └── data_suv_bi.csv     # Numeric features
+└── Cohort_B/
+    ├── data.csv
+    └── data_suv_bi.csv
+```
 
 Before running the workflow, ensure you have the following files prepared and know their locations:
 
@@ -73,6 +90,8 @@ You can inspect these default files using `system.file("files", "<filename>", pa
 
 **Define Your Paths (Replace Placeholders):**
 
+<details>
+<summary><strong>Click to expand: Comprehensive Path Setup Code</strong></summary>
 ```r
 # --- USER: Define paths to YOUR files and directories ---
 
@@ -127,11 +146,11 @@ pkg_literature_defs_file <- system.file("files", "literature_biomarkers.yaml", p
 # config <- yaml::read_yaml(custom_config_path)
 # If using package default config (preprocess_data will load it):
 # For other functions, we'll get it from preprocess_data output.
-# For now, let's assume preprocess_data handles config loading.
 
 ```
+</details>
 
-## Step 1: Loading Datasets
[…]
