### 2026-02-10 #9070 @isaac-6
@jromanowska Vignettes have been updated in the latest [commit](https://github.com/isaac-6/biodiscvr/commit/eaf7c0b4da031eceb1ffbd3b1242903168306220).

The synthetic‑data demo (`biodiscvr_synth`) has been reworked with clearer explanations, visible outputs, and a clickable table of contents.
The second vignette (`biodiscvr`) remains more code‑focused for users who prefer step‑by‑step guidance.

Thanks again for the guidance! Happy to adjust further if needed.
