### 2026-02-11 72f4c65
formatted the vignettes for style (cleaner lists)
