### 2026-02-12 6f4d410
Small wording fixes.
