+++ DESCRIPTION
@@ -1,9 +1,12 @@
 Package: biodiscvr
 Title: Biomarker Discovery Using Composite Value Ratios
-Version: 1.2.1
+Version: 1.3.0
 Date: 2025
-Authors@R: 
-    person("Isaac", "Llorente-Saguer", , "illorentes@gmail.com", role = c("aut", "cre"), comment = c(ORCID = "0000-0002-7612-5331"))
+Authors@R: c(
+    person("Isaac", "Llorente-Saguer", , "illorentes@gmail.com", role = c("aut", "cre"), comment = c(ORCID = "0000-0002-7612-5331")),
+    person("Neil P.", "Oxtoby", role = "aut", comment = c(ORCID = "0000-0003-0203-3909"),
+           email = "n.oxtoby@ucl.ac.uk")
+           )
 Description: Provides tools for discovering and evaluating biomarkers derived
     from ratios of composite feature sets (e.g., regional imaging measurements).
     Implements a Genetic Algorithm (GA) based search to optimize feature
