### 2026-02-12 #9070 @isaac-6
@jromanowska I completed the requested steps:
- Release: [v1.3.0: JOSS Review Completion Release](https://github.com/isaac-6/biodiscvr/tree/v1.3.0)
- Zenodo Archive DOI: [10.5281/zenodo.18621404](https://doi.org/10.5281/zenodo.18621404)
- Verification: The title and author list (including ORCIDs) in the Zenodo metadata match the paper. Both the repo and Zenodo have a GPL-3.0 license.

Thank you again for your guidance throughout this process!
