### 2026-02-12 #9070 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
