### 2026-02-12 8d0a818
Added DOI badge.
