### 2026-03-06 #9070 @rkurchin
👋 Hi, congrats on being nearly done! As AEiC for JOSS I will now help to process this submission for acceptance in JOSS (and apologies for the unusually long delay in getting this process started!). Below are some final checks, some of which may require your attention:
## Remaining points (entirely copyediting):
- [x] Make sure that capitalization is being `{Enforced}` where needed in your bibliography – e.g. for proper nouns R (the language) and Alzheimer's
- [x] citation at the end of second paragraph of Statement of need should be in parentheses (like the one at the start of the section)
- [x] be consistent about whether references to other section headers are italicized or not (see two references to Methods section)
- [x] bottom of page 3: would be good to have a sentence or two explaining the bullet list as it sort of comes out of nowhere as written currently – likely something as simple as "Important ideas and parameters in the equation above include:" or similar would suffice
- [x] maybe I'm just not fully understanding the notation (this is not my specific field of expertise) but why is the $s$ parameter needed if you take the absolute values of all the $f_i$'s?
- [x] I suggest renaming the "Citations" section to something more like "Prior Work" or similar, since it's a bit odd to have that header and then also "References" later on
