### 2026-03-06 3f62511
Fix formatting in bibliography entries (capitalisation)
Updated titles in bibliography entries to use proper formatting for 'Alzheimer's' and other terms.
