+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{llorente2024,
-  title={A data-driven framework for biomarker discovery applied to optimizing modern clinical and preclinical trials on Alzheimer’s disease},
+  title={A data-driven framework for biomarker discovery applied to optimizing modern clinical and preclinical trials on {Alzheimer's} disease},
   author={Llorente-Saguer, Isaac and Oxtoby, Neil P},
   journal={Brain Communications},
   volume={6},
@@ -11,9 +11,9 @@
 }
 
 @article{saguer2022composite,
-  title={Composite SUVR: A new method for boosting Alzheimer’s disease monitoring and diagnostic performance, applied to tau PET},
+  title={Composite {SUVR}: A new method for boosting {Alzheimer's} disease monitoring and diagnostic performance, applied to tau PET},
   author={Saguer, Isaac Llorente and Busche, Marc A and Oxtoby, Neil P},
-  journal={Alzheimer's \& Dementia},
+  journal={{Alzheimer's} \& Dementia},
   volume={18},
   pages={e063177},
   year={2022},
@@ -23,7 +23,7 @@
 
 
 @article{scrucca2016some,
-  title={On some extensions to GA package: hybrid optimisation, parallelisation and islands evolution},
+  title={On some extensions to {GA} package: hybrid optimisation, parallelisation and islands evolution},
   author={Scrucca, Luca},
   journal={arXiv preprint arXiv:1605.01931},
   year={2016},
@@ -42,9 +42,9 @@
 }
   
 @article{longpower,
-  title = {Power and Sample Size for Longitudinal Models in R-The longpower Package and Shiny App},
+  title = {Power and Sample Size for Longitudinal Models in {R}-The longpower Package and Shiny App},
   author = {Samuel Iddi and Michael C. Donohue},
-  journal = {R Journal},
+  journal = {{R} Journal},
   year = {2022},
   volume = {14},
   number = {1},
@@ -54,7 +54,7 @@
 }
 
 @article{llorente2026enhanced,
-  title={Enhanced monitoring of Alzheimer's disease brain atrophy using composite value ratios of volumes},
+  title={Enhanced monitoring of {Alzheimer's} disease brain atrophy using composite value ratios of volumes},
   author={Llorente-Saguer, Isaac and Oxtoby, Neil P},
   journal={Brain Communications},
   volume={8},
@@ -66,7 +66,7 @@
 }
 
 @article{llorente2025theta,
-  title={Theta, a multidimensional ratio biomarker applied to five amyloid beta peptides for investigations in familial Alzheimer’s disease},
+  title={Theta, a multidimensional ratio biomarker applied to five amyloid beta peptides for investigations in familial {Alzheimer's} disease},
   author={Llorente-Saguer, Isaac and Arber, Charles and Oxtoby, Neil P},
   journal={medRxiv},
   pages={2025--08},
