### 2026-03-06 d69d861
Fix citation format for references in paper.md
Inserted parentheses
