### 2026-03-06 b71272e
Enhance description of multicohort regularisation
Clarify the explanation of multicohort regularisation and its components in the global fitness function.
