### 2026-03-06 c635bfc
Update explanation for Directional Sign in fitness function
Clarify the condition for setting the directional sign based on individual cohort fitnesses.
