+++ paper/paper.md
@@ -155,7 +155,7 @@ The following elements capture the intuition behind the global fitness function 
 *   **Product-based Aggregation:** By using the product of fitnesses rather than the sum, the framework ensures that a biomarker must perform well across *all* cohorts. A low performance in any single cohort will heavily penalise the global score.
 *   **Reference Direction ($\mathbf{R}$):** This vector represents the desired balance of performance across cohorts. While it can be a vector of ones $[1, 1, \dots, 1]$, the framework defaults to the single best performance achieved per cohort, defining an adaptable performance ceiling.
 *   **Angular Penalty ($\cos^2\theta$):** This term penalises candidate biomarkers with high angular deviation from the reference, effectively enforcing consensus across cohorts (i.e., avoiding overfitting to a subset of cohorts).
-*   **Directional Sign ($s$):** To ensure biological validity, $s$ is set to $1$ only if all individual cohort fitnesses $f_i$ are positive. If any $f_i < 0$ (indicating a reversed effect direction), $s$ is set to $-1$.
+*   **Directional Sign ($s$):** To ensure biological validity, $s$ is set to $1$ only if all individual cohort fitnesses $f_i$ are positive. If any $f_i < 0$ (indicating a reversed effect direction), $s$ is set to $-1$. Otherwise, the product of negative individual fitnesses could cancel out the sign.  
 
 ### Optimisation Strategy
 The framework navigates the combinatorial space of feature ratios and groupings using an exploratory search algorithm, currently implemented via a Genetic Algorithm. The use of the directional sign $s$ creates a "hard barrier" in the search space, guiding the algorithm away from invalid solutions. This approach enables the discovery of complex, multi-feature biomarkers that are robust across heterogeneous datasets and would be computationally intractable to identify via exhaustive search.
