### 2026-03-06 a423735
Rename 'Citations' section to 'Prior Work'
Updated section title from 'Citations' to 'Prior Work' and adjusted related text.
