### 2026-03-06 #9070 @isaac-6
Hi Rachel @rkurchin , thank you for taking over this submission, and for the suggestions to improve the quality of this work.

Done:
- Make sure that capitalization is being {Enforced}
- Citation at the end of second paragraph of Statement of need should be in parentheses
- Be consistent about whether references to other section headers are italicized or not (see two references to Methods section)
- bottom of page 3: introductory sentence for the list.
- Sign of "s": if there are two cohorts that have a negative direction (and thus individual fitness), their multiplied fitness will be positive, and I wanted to avoid this. s = min(fitnesses) < 0, basically. I can understand it can be confusing, though.
-  I suggest renaming the "Citations" section to something more like "Prior Work"
