### 2026-03-06 78e5c5b
Revise citation format
Updated citation format in the discussion of `biodiscvr` features.
