### 2026-03-06 #9070 @rkurchin
I think the Markdown for the bullet list on page 3 got mangled, could you take a look and fix that?

On the sign of $s$ thing, I still don't understand, because in the equation you're taking the absolute value, so the product should always be positive, no?
