+++ paper/paper.md
@@ -69,7 +69,7 @@ to discover high-performing CVRs tailored to specific clinical or research objec
 such as maximising group separation or minimising sample size requirements ([@saguer2022composite; @llorente2024; @llorente2026enhanced]).
 
 A novel contribution of this implementation is the introduction 
-of **multicohort regularisation** (see *Methods*). Traditionally, biomarker discovery 
+of **multi-cohort regularisation** (see *Methods*). Traditionally, biomarker discovery 
 is performed on single datasets or pooled data. Pooling often requires intensive 
 data harmonisation to mitigate batch effects and can lead to models that overfit 
 to the most prevalent data structure. `biodiscvr` advances the CVR framework by 
@@ -117,7 +117,7 @@ The core discovery process is managed by `run_experiments()`, which manages sing
 or multi-cohort optimisation runs. This function interfaces with the `GA` package to navigate the 
 high-dimensional feature space. In each generation of the search, candidate biomarkers are 
 passed to an internal evaluation loop that fits linear mixed-effects models (via `lme4`) and 
-calculates clinical utility metrics (via `lmmpower`). The multicohort consensus logic (see 
+calculates clinical utility metrics (via `lmmpower`). The multi-cohort consensus logic (see 
 *Methods*) is then applied to aggregate these metrics into a single fitness score.
 
 ### 4. Evaluation and Parsimony Refinement
@@ -144,14 +144,14 @@ For each cohort $i$, the framework fits a linear mixed-effects model to the log-
 
 These are combined into a scalar fitness value $f_i = t/N$. To avoid a state where the algorithm cannot improve sample size without sacrificing group separation (Pareto frontier), the $t$-statistic is truncated at a saturation point (defaulting to 2.6, or $p \approx 0.005$). This allows the optimiser to prioritise the reduction of $N$ once a sufficient threshold of statistical significance is reached.
 
-### Multicohort Regularisation
+### Multi-cohort Regularisation
 To identify biomarkers that generalise across populations without merging datasets (thus avoiding batch effects), `biodiscvr` treats the fitness scores from $K$ cohorts as a vector $\mathbf{F} = [f_1, f_2, \dots, f_K]$. The global fitness $G$ is calculated using the product of absolute individual fitnesses, an angular penalty, and a directional sign:
 
 $$G = s \cdot \left( \prod_{i=1}^{K} |f_i| \right) \cdot \cos^2(\theta)$$
 
 Where $\theta$ [@llorente2025theta] is the angle between the fitness vector $\mathbf{F}$ and a **reference direction vector** $\mathbf{R}$, and $s$ represents the directional consistency.
 
-The following elements capture the intuition behind the global fitness function and how it enforces multicohort consistency:
+The following elements capture the intuition behind the global fitness function and how it enforces multi-cohort consistency:
 *   **Product-based Aggregation:** By using the product of fitnesses rather than the sum, the framework ensures that a biomarker must perform well across *all* cohorts. A low performance in any single cohort will heavily penalise the global score.
 *   **Reference Direction ($\mathbf{R}$):** This vector represents the desired balance of performance across cohorts. While it can be a vector of ones $[1, 1, \dots, 1]$, the framework defaults to the single best performance achieved per cohort, defining an adaptable performance ceiling.
 *   **Angular Penalty ($\cos^2\theta$):** This term penalises candidate biomarkers with high angular deviation from the reference, effectively enforcing consensus across cohorts (i.e., avoiding overfitting to a subset of cohorts).
@@ -163,7 +163,7 @@ The framework navigates the combinatorial space of feature ratios and groupings 
 
 # Prior Work
     
[…]
