### 2026-03-06 #9070 @isaac-6
**RE: fitness.**
The directional sign term s is necessary to penalise candidate biomarkers that produce negative individual fitness values. Without this term, a biomarker that performs poorly in one or more cohorts could still appear acceptable when the global fitness is computed. For example, suppose cohort A yields an individual fitness of -0.5 and cohort B yields -0.3. Their product is 0.15, which is numerically identical to the product of two positive fitness values 0.5 and 0.3. This would incorrectly suggest that the biomarker performs well across cohorts.

To avoid this ambiguity, the framework takes the absolute value of each individual fitness before aggregation, ensuring the product reflects magnitude only. The directional sign s is then applied to encode whether all cohorts agree in effect direction. If _any_ cohort produces a negative fitness value, s=-1, making the global fitness explicitly unfavourable. This ensures that biomarkers with inconsistent or biologically implausible directionality are penalised appropriately.

This helps the convergence of the genetic algorithm.

**RE: list.**
Thank you! I forgot it renders differently that what I see. Hopefully fixed for good.

---
Please let me know if there’s anything else you’d like me to address. Thank you!
