### 2026-03-06 16aa402
Fix citation formatting in paper.md
